# swagger.yaml
openapi: 3.0.0
info:
    title: Jobs-API
    contact: {}
    version: "1.0"
servers:
    - url: http://localhost:3000/api/v1
      variables: {}
paths:
    /auth/register:
        post:
            tags:
                - Auth
            summary: Register User
            operationId: RegisterUser
            parameters: []
            requestBody:
                description: ""
                content:
                    application/json:
                        schema:
                            allOf:
                                - $ref: "#/components/schemas/RegisterUserRequest"
                                - example:
                                      name: test
                                      email: test@test.com
                                      password: secret
                        example:
                            name: test
                            email: test@test.com
                            password: secret
                required: true
            responses:
                "200":
                    description: ""
                    headers: {}
            deprecated: false
            security: []
    /auth/login:
        post:
            tags:
                - Auth
            summary: Login User
            operationId: LoginUser
            parameters: []
            requestBody:
                description: ""
                content:
                    application/json:
                        schema:
                            allOf:
                                - $ref: "#/components/schemas/LoginUserRequest"
                                - example:
                                      email: bitch@test.com
                                      password: secret
                        example:
                            email: bitch@test.com
                            password: secret
                required: true
            responses:
                "200":
                    description: ""
                    headers: {}
            deprecated: false
            security: []
    /jobs:
        post:
            tags:
                - Jobs
            summary: Create Job
            operationId: CreateJob
            parameters: []
            requestBody:
                description: ""
                content:
                    application/json:
                        schema:
                            allOf:
                                - $ref: "#/components/schemas/CreateJobRequest"
                                - example:
                                      company: Bitchology
                                      position: Software Engineer
                        example:
                            company: Bitchology
                            position: Software Engineer
                required: true
            responses:
                "200":
                    description: ""
                    headers: {}
            deprecated: false
            security:
                - bearer: []
        get:
            tags:
                - Jobs
            summary: Get All Jobs
            operationId: GetAllJobs
            parameters: []
            responses:
                "200":
                    description: ""
                    headers: {}
            deprecated: false
            security:
                - bearer: []
    /jobs/{id}:
        parameters:
            - in: path
              name: id
              schema:
                  type: string
              required: true
              description: The job ID
        get:
            tags:
                - Jobs
            summary: Get Job
            operationId: GetJob
            parameters: []
            responses:
                "200":
                    description: ""
                    headers: {}
            deprecated: false
            security:
                - bearer: []
        patch:
            tags:
                - Jobs
            summary: Update Job
            operationId: UpdateJob
            parameters: []
            requestBody:
                description: ""
                content:
                    application/json:
                        schema:
                            allOf:
                                - $ref: "#/components/schemas/UpdateJobRequest"
                                - example:
                                      company: New bitch Company
                                      position: New Position
                        example:
                            company: New bitch Company
                            position: New Position
                required: true
            responses:
                "200":
                    description: ""
                    headers: {}
            deprecated: false
            security:
                - bearer: []
        delete:
            tags:
                - Jobs
            summary: Delete Job
            operationId: DeleteJob
            parameters: []
            responses:
                "200":
                    description: ""
                    headers: {}
            deprecated: false
            security:
                - bearer: []
components:
    schemas:
        RegisterUserRequest:
            title: RegisterUserRequest
            required:
                - name
            type: object
            properties:
                name:
                    type: string
            example:
                name: bitch
        LoginUserRequest:
            title: LoginUserRequest
            required:
                - email
                - password
            type: object
            properties:
                email:
                    type: string
                password:
                    type: string
            example:
                email: bitch@test.com
                password: secret
        CreateJobRequest:
            title: CreateJobRequest
            required:
                - company
                - position
            type: object
            properties:
                company:
                    type: string
                position:
                    type: string
            example:
                company: Bitchology
                position: Software Engineer
        UpdateJobRequest:
            title: UpdateJobRequest
            required:
                - company
                - position
            type: object
            properties:
                company:
                    type: string
                position:
                    type: string
            example:
                company: New bitch Company
                position: New Position
    securitySchemes:
        bearer:
            type: http
            scheme: bearer
security: []
tags:
    - name: Auth
      description: ""
    - name: Jobs
      description: ""